recommendations from python API (+ related changes and tests)

+++ src/main/java/com/gamesUP/gamesUP/DTO/User/AvisDTO.java
@@ -7,6 +7,6 @@ public class AvisDTO {
     private Long userId;
     private Long gameId;
     private String commentaire;
-    private Integer note;
+    private Float note;
 }
